+++ crates/tiny-loop/examples/bind_tools.rs
@@ -1,7 +1,11 @@
 mod common;
 
 use common::run_cli_loop;
-use std::{collections::HashMap, sync::Arc};
+use std::{
+    collections::HashMap,
+    io::{Write, stdout},
+    sync::Arc,
+};
 use tiny_loop::{Agent, llm::OpenAIProvider, tool::tool};
 use tokio::sync::Mutex;
 
@@ -66,7 +70,11 @@ async fn main() {
     let llm = OpenAIProvider::new()
         .api_key(api_key)
         .base_url("https://openrouter.ai/api/v1")
-        .model("google/gemini-3-flash-preview");
+        .model("google/gemini-3-flash-preview")
+        .stream_callback(|chunk| {
+            print!("{}", chunk);
+            stdout().flush().unwrap();
+        });
 
     let mut data = HashMap::new();
     data.insert("name".to_string(), "Alice".to_string());